pyproject.toml
[project]
name = "mcp-graph-mailer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.3",
    "httpx>=0.28.1",
    "msal>=1.34.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.13",
]
